eco32: head: implement tlbmiss handler
This is the fast path of the tlbmiss, anything that isn't
handled by this will automatically be handled by the page fault
handler.
We make use of two fixed registers, $1 is the asm temp register,
gas will use this to insert ldhi,ori sequences on large
immediates.
It will be inserted on the 'ori $1,$0,current_pgd', but apart
from a wasted instruction, it will be ok.

Using $3 is a bit more controversial, but until gcc start to
make use of it, it should be fine.

The handler is not tested at all, so there might be dragons...

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>