eco32: use $26 as scratch reg instead of $3 in exceptions
I initially picked $3 (return high value) since that is unused
by gcc and I had some thought about saving $26 (that is a proper
OS reserved register) for something else (got pointer for instance).
I don't think that's motivated though, so this exchanges all usage
of $3 as a scratch reg with $26.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>