eco32: process: copy PSW_IC to PSW_IP in start_thread
Since the regs in pt_regs for the thread that will be started is the post
exception regs, we have to put the flags in the 'previous' bit flags
so they will be loaded on the return from exception.
The UP flag was already handled correctly, this fixes so that the
global interrupt flag also is handled correctly.

Signed-off-by: Stefan Kristiansson <stefan.kristiansson@saunalahti.fi>